+++ server/web/assets/js/subscriptions.js
@@ -12,7 +12,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Incorrect Data";
-				$("#subscription_result").html("<p>" + message + "</p>");
+				$("#subscription_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -21,8 +21,11 @@ $(document).ready(function () {
 	$("a[id^='delete_subscription']").click(function () {
 		var id = this.id.replace("delete_subscription_", "");
 		var customerId = this.dataset.customerId;
+		if (!confirm(`Are you sure you want to delete ${this.title.replace("Delete ", "")} subscription and all its licenses?`)) {
+			return
+		}
 		$.ajax({
-			url: `/api/customer${customerId}/subscription/${id}`,
+			url: `/api/customer/${customerId}/subscription/${id}`,
 			type: "DELETE",
 			success: function () {
 				location.reload();
@@ -31,7 +34,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Cannot delete subscription";
-				$("#subscriptions_result").html("<p>" + message + "</p>");
+				$("#subscriptions_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -55,7 +58,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Cannot modify subscription";
-				$("#subscriptions_result").html("<p>" + message + "</p>");
+				$("#subscriptions_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -64,7 +67,7 @@ $(document).ready(function () {
     $("a[id^='toggle_license_']").click(function () {
 		var id = this.id.replace("toggle_license_", "");
 		$.ajax({
-			url: `/api/key/${id}`,
+			url: `/api/license/${id}`,
 			type: "PATCH",
 			data: {
                 status: this.dataset.status == "true" ? "" : "on", 
@@ -76,7 +79,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Cannot modify license";
-				$("#subscriptions_result").html("<p>" + message + "</p>");
+				$("#subscriptions_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -86,7 +89,7 @@ $(document).ready(function () {
 		var id = this.id.replace("delete_license_", "");
 		var customerId = this.dataset.customerId;
 		$.ajax({
-			url: `/api/key/${id}`,
+			url: `/api/license/${id}`,
 			type: "DELETE",
 			success: function () {
 				location.reload();
@@ -95,7 +98,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Cannot delete subscription";
-				$("#subscriptions_result").html("<p>" + message + "</p>");
+				$("#subscriptions_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
