+++ server/web/assets/js/tariff.js
@@ -5,13 +5,13 @@ $(document).ready(function () {
 			type: "POST",
 			data: $("#tariff_form").serialize(),
 			success: function () {
-				location.reload();;
+				location.replace("/admin/tariffs");
 			},
 			error: function (xhr, status, error) {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Incorrect Data";
-				$("#tariff_result").html("<p>" + message + "</p>");
+				$("#tariff_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -32,7 +32,7 @@ $(document).ready(function () {
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != ""
 						? err.msg
 						: "Incorrect Data";
-				$("#tariff_result").html("<p>" + message + "</p>");
+				$("#tariff_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
@@ -40,6 +40,9 @@ $(document).ready(function () {
 
 	$("a[id^='delete_tariff']").click(function () {
 		var id = this.id.replace("delete_tariff_", "");
+		if (!confirm(`Are you sure you want to delete ${this.title.replace("Delete ", "")} plan and all its subscriptions?`)) {
+			return
+		}
 		$.ajax({
 			url: "/api/tariff/" + id,
 			type: "DELETE",
@@ -50,7 +53,7 @@ $(document).ready(function () {
 				var err = eval("(" + xhr.responseText + ")");
 				var message =
 					typeof err.msg !== "undefined" && err.msg != null && err.msg != "" ? err.msg : "Cannot delete plan";
-				$("#tariffs_result").html("<p>" + message + "</p>");
+				$("#tariffs_result").html("<p><b>" + message + "</b></p>");
 			},
 		});
 		return false;
